swift: Add early platform support for x86_64-unknown-haiku
This commit allows swift to be built on Haiku, excluding testing support.
At the moment, the compiler and runtime is non-functional, but all other
developer tools work as expected.

As for the frontend (swiftc) it can parse swift sources and emit SIL/IR
but fails to generate a executable. (It polls forever when compiling a hello world.)